Extract static methods, simplify
